update oh-my-zsh for multi platfrom

+++ oh-my-zsh/uninstall.sh
@@ -7,19 +7,28 @@ ZSH_PLUGINS="$CUR_DIR/plugins.zsh"
 SOURCE_ZSH_PLUGINS="source $ZSH_PLUGINS"
 SHELL_RC="$HOME/.zshrc"
 
-# 1. Remove zsh plugins by brew
-brew remove \
-  zsh-autosuggestions \
-  zsh-syntax-highlighting \
-  zsh-completions
+[ ! -n "$ZSH_VERSION" ] || { echo "You are not using Zsh"; exit 1; }
 
-# 2. Remove shell profile for brew zsh plugins
+# 1. Remove zsh custom plugins
+if command -v brew &>/dev/null; then
+    brew remove \
+        zsh-autosuggestions \
+        zsh-syntax-highlighting \
+        zsh-completions
+else
+    ZSH_PLUGIN_PREFIX="${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/plugins"
+    rm -rf "${ZSH_PLUGIN_PREFIX}/zsh-completions"
+    rm -rf "${ZSH_PLUGIN_PREFIX}/zsh-autosuggestions"
+    rm -rf "${ZSH_PLUGIN_PREFIX}/zsh-syntax-highlighting"
+fi
+
+# 2. Remove shell profile for zsh custom plugins
 if grep -q "$SOURCE_ZSH_PLUGINS" "$SHELL_RC"; then
     grep -vF "$SOURCE_ZSH_PLUGINS" "$SHELL_RC" > "$SHELL_RC.tmp" && mv "$SHELL_RC.tmp" "$SHELL_RC"
-    echo "Zsh plugins config successfully removed from $SHELL_RC"
+    echo "ZSH plugins config successfully removed from $SHELL_RC"
     source $SHELL_RC
 else
-    echo "No Zsh plugins config found in $SHELL_RC. No changes made."
+    echo "No ZSH plugins config found in $SHELL_RC. No changes made."
     exit
 fi
